# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HaloLight OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/ai-knowledge:
    get:
      operationId: listAIKnowledgeDocs
      tags: [ai-knowledge]
      summary: List AI knowledge documents
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: aiActive
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: AI knowledge document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIKnowledgeDocList"
    post:
      operationId: createAIKnowledgeDoc
      tags: [ai-knowledge]
      summary: Create an AI knowledge document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAIKnowledgeDocInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIKnowledgeDoc"
  /admin/ai-knowledge/{id}:
    get:
      operationId: getAIKnowledgeDoc
      tags: [ai-knowledge]
      summary: Get AI knowledge document with chunks
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: AI knowledge document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIKnowledgeDocDetail"
    put:
      operationId: updateAIKnowledgeDoc
      tags: [ai-knowledge]
      summary: Update an AI knowledge document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAIKnowledgeDocInput"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIKnowledgeDoc"
    delete:
      operationId: deleteAIKnowledgeDoc
      tags: [ai-knowledge]
      summary: Delete an AI knowledge document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /admin/ai-knowledge/{id}/reindex:
    post:
      operationId: reindexAIKnowledgeDoc
      tags: [ai-knowledge]
      summary: Chunk and reindex a knowledge document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reindex result
          content:
            application/json:
              schema:
                type: object
                required: [documentId, chunksCreated]
                properties:
                  documentId: { type: string }
                  chunksCreated: { type: integer }

  # ─── Admin Academy ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AIKnowledgeDoc:
      type: object
      required:
        [
          id,
          title,
          language,
          category,
          content,
          tags,
          aiActive,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string }
        title: { type: string }
        language: { type: string }
        category:
          {
            type: string,
            enum:
              [
                faq,
                troubleshooting,
                printer_manual,
                camera_manual,
                software_guide,
                business_guide,
                pricing_guide,
                event_guide,
                product_guide,
                academy_lesson,
                support_article,
              ],
          }
        productModel: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        content: { type: string }
        tags: { type: array, items: { type: string } }
        aiActive: { type: boolean }
        lastIndexedAt: { type: string, nullable: true }
        status: { type: string, enum: [draft, indexed, needs_review, archived] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AIKnowledgeChunk:
      type: object
      required: [id, documentId, content, chunkIndex]
      properties:
        id: { type: string }
        documentId: { type: string }
        content: { type: string }
        chunkIndex: { type: integer }
        metadata: { type: object }
        createdAt: { type: string }
    AIKnowledgeDocDetail:
      type: object
      required:
        [
          id,
          title,
          language,
          category,
          content,
          tags,
          aiActive,
          status,
          createdAt,
          updatedAt,
          chunks,
          tagList,
        ]
      properties:
        id: { type: string }
        title: { type: string }
        language: { type: string }
        category: { type: string }
        productModel: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        content: { type: string }
        tags: { type: array, items: { type: string } }
        aiActive: { type: boolean }
        lastIndexedAt: { type: string, nullable: true }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/AIKnowledgeChunk"
        tagList:
          type: array
          items: { type: string }
    AIKnowledgeDocList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AIKnowledgeDoc"
        total: { type: integer }
    CreateAIKnowledgeDocInput:
      type: object
      required: [title, category]
      properties:
        title: { type: string }
        language: { type: string }
        category:
          {
            type: string,
            enum:
              [
                faq,
                troubleshooting,
                printer_manual,
                camera_manual,
                software_guide,
                business_guide,
                pricing_guide,
                event_guide,
                product_guide,
                academy_lesson,
                support_article,
              ],
          }
        productModel: { type: string }
        sourceUrl: { type: string }
        content: { type: string }
        tags: { type: array, items: { type: string } }
        aiActive: { type: boolean }
        status: { type: string, enum: [draft, indexed, needs_review, archived] }
    UpdateAIKnowledgeDocInput:
      type: object
      properties:
        title: { type: string }
        language: { type: string }
        category:
          {
            type: string,
            enum:
              [
                faq,
                troubleshooting,
                printer_manual,
                camera_manual,
                software_guide,
                business_guide,
                pricing_guide,
                event_guide,
                product_guide,
                academy_lesson,
                support_article,
              ],
          }
        productModel: { type: string }
        sourceUrl: { type: string }
        content: { type: string }
        tags: { type: array, items: { type: string } }
        aiActive: { type: boolean }
        status: { type: string, enum: [draft, indexed, needs_review, archived] }
...